Refresh README, examples and changelog with recent changes
This PR refreshes README, examples of our and changelog to better
emphasise recent changes in the validation/open API of ours
clarifying read-only properties in our API.

Specifically DAGRun dag_id has not been marked as read-only so far
which was a bug, and some of our "dev" examples set the read only
properties, which could mislead our users.

We've recently clarified that those properties are read-only and
we are now updating the API documentation and examples to show that
the good way of using related APIs.

Also the examples have been refrehsed and modernized a bit as well
as the dev example has been synchronized automatically (via
pre-commit) to make sure it is the same in the README and in dev
where we are using it to run the tests with the API.

Minimum version of Python have been set to 3.7 and CHANGELOG was
updated to reflect all the changes, as well as it has been
exposed in Project_URL of the package in PyPI.

Link in README have been changed to URL so that they are properly
rendered in README in `pypi` package documentation (currently
following the links to other filess in the repo there fails.