Housekeeping of Python client repository
After working on reproducible package proparation for Python client
it's been found that the Python client repository had some extra
code added and docs and tests were placed in a wrong directory.

As of https://github.com/apache/airflow-client-python/pull/93, we
have duplicated airflow_client code:

* airflow_client.client - this is the one that should be used
* airflow_client.airflow_client - this is the one that is added
  accidentally (and it is a copy of the `client` one)

Also `docs` and `tests` are generated in "airflow_client" package,
but this is wrong, both docs and tests should be at the top-level
of the repository.

This PR:

* deletes the extra code
* moves "docs" and "test" folders to the top-level of the project.